MERO-821: more fix for use after free for cp
An accumulator copy packet is also finalised as a normal copy packet
(->co_free()), just that it is not freed as a normal copy packet,
because it is part of struct m0_sns_cm_repair_ag_failure_ctx, so the
memory for it is freed as part of the failure context finalisation
during aggregation group finalisation.
So we need to invoke ->co_free() before ag->cag_ops->cago_fini(ag) as
the later can free the memory allocated for the accumulator copy packet.

commit 3ef056ff1b4da6615e1e0273cb6644c8d00371b4 has a fix, but it
doesn't correct.

Change-Id: I640194ae9c0fdf21c97ec94991303c4d0ba2084e