MERO-857: rc is used unintialised
A failures context represents a failed data unit in an aggregation
group and holds information about the corresponding spare such as
its location and cob id. Number of failure contexts are equal to
number of failures in an aggregation group. There's a possibility
that a particular failure context is not used, see repair_ag_failure_ctxs_setup()
for more details. In funtion m0_sns_cm_repair_ag_setup() rc is unintialised,
and is changed only if a failure context is in-use. Thus if a failure
context remains unused (especially the 0th one), rc remains unchanged
and may have garbage.

Change-Id: I90c95b3b28be5213cb852e7cf6198bd12ffbe7fb