balloc: fix allocation struggles as more data is allocated
After analysing the log and the code it looks like the
situation is following. All groups have single fragment
which start is not aligned with requested size (for example,
the fragments starts from 0x00xxxx80 while the requested
size is 0x100). In this case the balloc_find_extent_buddy()
will not find such buddy, because it requires the fragment
start to be aligned with multiple of requested size. So
balloc goes over all the groups before it eventually gives
up and allocates from some of the group. And now is the most
ugly thing - when the new allocation recorded - it does not
insert the new size-aligned record to the group, but instead
- just updates the existing unaligned one! As result - there
is no chance for a new allocations of the same size to find
and use this group. As result, every single new allocation
goes over all groups again and again.

Now, we try to insert the new record into the group which is
aligned with allocated size. So the next time allocation of
the same (or smaller) size will immediately find this group.

Reviewed-by: Hua Huang <hua.huang@seagate.com>
Reviewed-by: Nikita Danilov <nikita.danilov@seagate.com>

References: https://jira.xyratex.com/browse/MERO-877

Change-Id: I1c8fcf79a29743bdac1b04db7b4d8c8adaddbbb6