balloc: avoid btree deletions on alloc
balloc was designed to use btree in such a way that it
deleted (very heavy operation) and inserted (also heavy
operation) the records for every single allocation. Now, it
just updates the existing btree record that is very easy
operation.

balloc uses m0_ext to record the free areas. ext->e_start
was used as a key and ext->e_end was used as a value in
btree record. Thus, when the new value is allocated from the
start region, we needed to update the btree key, i.e. we
needed to delete and insert the new record with updated key.
Now, we use ext->e_end as a key, so on allocation we only
update the value of the record - that is very easy btree
operation.

Reviewed-by: Hua Huang <hua.huang@seagate.com>

References: https://jira.xyratex.com/browse/MERO-882

Change-Id: Ibf36217790808109e7184b1e3da4fdf4bfc98ecd