rm: Check cached credits first against incoming request
In some cases it is useful to check cached credits in
incoming_check_with() and only after that check held credits.

For example, if incoming revoke request has matching credit in both
cached and held credits lists, it's better to use cached credit, so
there is no need to call cro_conflict() callback for held credit and
wait for user to put this held credit.

* Iterate over owned credits in reverse order in incoming_check_with().

* Add assertion in owner_invariant() checking that no any pair of
  owned credit conflicts. It assures that no cached credit conflicts
  with held credit, so behaviour of incoming_check_with() for requests
  with RIF_LOCAL_WAIT didn't change.

Change-Id: Ia84471d87266ad6009462518f930c19cf8b05b14