be/engine: place only most recently captured regions
It was possible to capture the same region in tx1 first
and then the same region in tx2 and place it after
the latter that led to BE segment data corruption.

Now we introduce monotonically increasing generation
index to regions, and only the most recently captured ones
are placed.

Reviewed-by: Andriy Tkachuk <andriy.tkachuk@seagate.com>
Reviewed-by: Yuriy Umanets <yuriy.umanets@seagate.com>
Reviewed-by: Anatoliy Bilenko <anatoliy.bilenko@seagate.com>

References: https://jira.xyratex.com/browse/MERO-915

Change-Id: I95f988c80c0f3e8704b4fd2ee687b1f24bb42c41