MERO-1033 Support for multiple pool-versions in SNS repair
SNS and pool-utility were updating device state in the "global" pool-machine
present in the request handler. Since each file can have a different
pool-version associated with it, and as each pool-version has its own
pool-machine, updating only the global pool-machine does not suffice.

This issue is addressed by updating all pool-machines to which a device
belongs, whenever its state gets updated in the global pool-machine.

Couple of fixes in m0t1fs have also been made. These
are relevant to MERO-1019 (parity verifications fails on healthy parity groups
in degraded mode read) and MERO-588 (spares are not being written to during
writes post repair). Currently an issue comes with sns-cm-repair-ut, and
so it has been blocked. MERO-1096 is created for the same.

Change-Id: I7b676254994b2a5772ccc568d249a78e7c1c2435