m0mount: updated after confc.HA change
After commit 5a37b98, mkfs tries to connect to confd, so a local
endpoint is provided it.
During m0d initialisation, it tries to connect to all other endpoints,
so make sure that mkfs does not share the endpoint of m0d on that node.
Confd and other service m0ds are run in between mkfs's, which may connect
to mkfs instead of m0ds.
Also a separate HA m0d is added as is needed after confc.HA change.

Change-Id: I9396b3d6feb6f7048d2c6152cf89dffe3c422370