MERO-452 SNS rebalance fails with -EIO on online device
When the disk is powered OFF and ON the disk device file (under /dev)
gets changed, so the device file needs to be reopened to access it.
This is the fix for an i/o error when such scenario occur during Mero
runtime. Since fd is stored in stob, as a fix, update fd by destroying
and creating the stob again when the disk is back online.

Change-Id: Ibf40bc5fbec7fe467f45d3efe36ad993720d6938