build: fix mero.rpm deps
Issue:

  [root@lt1-2b-cc1 ~]# rpm -i mero-0.1.0-3.10.0_123.9.3.el7.x86_64.gitb811450.el7.centos.x86_64.rpm error: Failed dependencies:
    libgalois.so.debug()(64bit) is needed by mero-0.1.0-1_3.10.0_123.9.3.el7.x86_64.gitb811450.el7.centos.x86_64
    libmero-ut.so.0()(64bit) is needed by mero-0.1.0-1_3.10.0_123.9.3.el7.x86_64.gitb811450.el7.centos.x86_64
    lustre-modules is needed by mero-0.1.0-1_3.10.0_123.9.3.el7.x86_64.gitb811450.el7.centos.x86_64
    perl(autodie) is needed by mero-0.1.0-1_3.10.0_123.9.3.el7.x86_64.gitb811450.el7.centos.x86_64

It resides deeply, somewhere inside rpmbuild scripts. I was not able to find a
proper solution to it. For some reason rpmbuild injects superfluous dependencies
on *.debug files (and other fluff) into mero.rpm.

As a workaround, generation of mero-debuginfo.rpm is disabled by default when
rpm are built locally with make rpms. In this case mero.rpm will not have
unsatisfied dependencies on *.debug files. In case debuinfo rpm might be needed,
it can be built like this: make rpms RPMBUILD_FLAGS='--with debuginfo'.

Change-Id: I41fd3322dbc58125c730d8f96815f2fb6048ba9f