MERO-1053 ios: File exist error
In non-oostore mode directly writing to the file sometimes fails with
EEXIST error. This error is occuring when pool version mismatch occurs
during write. This mismatch of pver is due to creation of io cobs during
setattr without valid pver.

This is fixed by sending valid pool version during setattr.

Also when pool version mismatch occurs during write, it should
recreate the io cob with new pool version, but is failing with
EEIXST error. This is found to be due to different cob index used
during io cob recreation than the one used during setattr.
This is fixed by using cob device as index in m0t1fs so that both
cob nsrec is same in cob_foms.c and io_foms.c during cob creation.

A test is also added in ST for directly writing to the file,
without creating it before.

Change-Id: Iee4576022e75c1731a8433e5a28a5133672b962e