MERO-1205 initscripts: Support parallel mkfs when external conf is used
As part of commit b0b2c200 (MERO-1168) during mkfs a local private endpoint
is used for confd to guard ioservices from other nodes to connect to
the confd which is not active (May be due to mkfs on confd node is finished).

Similarly when external configuration is used, a separate copy of conf is
made with local "private" conf endpoint and is used during mkfs on confd node.

Change-Id: Ifa5193d3114e5f5a2c441a4ee738f8e5e6881200