MERO-1148 Fix oostore mode ST failure with E2BIG
In m0t1fs_cob_getattr(), lid and pver are updated from getattr reply.
But layout rebuild is being skipped due to wrong if condition.
This happened during commit edf6037 OSAINT_mero-517.

This is fixed by updating the fields as well as layout inside a single
if condition.
Also found that pver is not updated after getattr in non-oostore mode
and is fixed.
M0_COB_PVER is added to verify pver cob attribute.
In ST, remount test case is added.

Change-Id: If7eb885b577e4cd57a81d7102cf0d89f4a314bd3