MERO-1244 confc: Repost a different confc request item in case of EAGAIN
When -EAGAIN is received (see commit 093ea23), confc re-posts rpc item.
The problem is that the rpc item does not differ from the one previously
sent, and confc receives the same (cached) reply

This is due to xid not being assigned on sending side as
m0_rpc_item::nr_sent is always greater than one.

Fix this problem by finalising the current rpc item and moving
back to the S_CHECK state, which posts the new rpc item.

conf_obj_status_reset() is added to reset the status of the object in
the request item.
m0_conf_cache_inquire() is removed as it returns the first
M0_CS_LOADING object, which is not needed.

Change-Id: I734c7f04aa9ebe842cfc2329a5cd7d361b65677a