MERO-1288 m0t1fs: Fix for setattr crash during iozone run
iozone calls close() after deleting the file, which calls flush()
in which file attributes are set for the deleted file, which is the
cause of iozone crash.

This is fixed by checking inode nlinks and setting the attributes only when
inode link count greater than zero.
Also in oostore mode after deleting the file inode link count is decremented
and inode is marked dirty for cleanup by the system.

Change-Id: I9a53c9164cf397d1a91407ff689e9fa75ca2a9af