MERO-1292 be/alloc: performance issues fixed
The problem was with growing number of small
free chunks in the beginning of free list which
were created on allocations, so each further
allocation should traverse all of them to find
the big one.

Now, all small free chunks go to the tail of
the free list and don't affect on future
allocations performance. This fixes the issue
MERO-1292 (Performance drops to half after
writing 12TB of data).

Reviewed-by: Maxim Medved <max.medved@seagate.com>

Change-Id: I6fe5996abae7a27ab6f8abe0f5b6260cbc8d7d94