MERO-1400 m0t1fs stuck during session cancellation
When rpc_post failed for setattr, semaphore down is
not called as value of 'i' is not incremented,
due to which semaphore is finalised before all the
callbacks are called, in which semaphore up is stuck.

This is observed during cancellation of md cob setattr
when md redundancy is greater than one. Also when CROW
flag is not set, this can occur in other cases too.

- M0_CONF_DISK_TYPE is added in ha note invariant.
- On tolerance check failure don't add pver in the list.
- session cancel/restore is added in service handler.
- On process failure event session is cancelled.
- conf file system params could be null.

As inode contains pver and lid, using getfattr these
values are printed.
Note: inode should be valid before use of getfattr.

Change-Id: I8fd69a63550cc426a9510b8aa0d17a30fd735a80