MERO-901 : ST for mdpools
This test covers the sanity of file operations whenever one of the
mds is down (because some failure domain at or above the level of
controllers, is down).
Most of the coverage for this ST has been covered in "pool version
assignment" st. In the original ST, the md redundancy was set to 1.
However, for correctly  testing md pools, the md redundancy needs to be
greater than 1 and this and related modifications have been done to the ST.
Also, the value of "pver" of a file is printed using the getfattr
command as and when needed. Added checks on  pool version ids after
the pool version switches.

Change-Id: I899be5a885d1cfa05b3c6dd0fdef257ea77e40a2