MERO-1423: sns-abort-test failure: nlx_buffer_invariant(nb)) at nlx_xo_buf_deregister()
Problem:
During aggregation group initialisation, m0_sns_ir is initialised (incremental recovery
context is initialised), during this initialisation the aggregation group accumulator
buffers are registered with sns incremental recovery module. If the unit size > max buffer
segment size (256) then multiple net_buffers are used per copy packet. In this case the
bufvecs of multiple buffers in the accumulator copy packet are merged with the buffer at the
head of the buffers list in the copy packet as a requirement of sns_ir module.
Later before posting this accumulator copy packet on network, this bufvec is split.
During abort it is possible that frozen aggregation groups get pruned, as part of pruning
these accumulator copy packets (which never got posted) buffers are released but without
splitting the merged bufvec. This results in the above invariant failure.

Fix:
Split the merged buffer from accumulator copy packet in-case the accumulator copy
packet is not posted on network.

Change-Id: Ic36a44dd9030f6e8ad30e5dff17b26e81e8ef2bc