/bin/bash -> /usr/bin/env bash
This is to allow compiling and running Mero in environments where bash
is located in a non-standard place, such as NixOS. Note that this
pattern is already used in the sources for calling python scripts.

There are two awk scripts (scripts/bloatmeter, utils/m0layout-from-addb)
where this replacement could not be applied because awk requires the -f
option to read the scripts and shebang lines can only have a single
argument.

Change-Id: I71399cf7c2ce51a95bb5cc8d6ad49003fbadeaf8