MERO-1534: m0_rpc_packet_decode_using_cursor() error handling panic fixed
There was problem in error handling path of m0_rpc_packet_decode_using_cursor()
routine which led to panic in case of corrupted rpc packet. The panic was caused
by fop_put_lock() which requires rpc_mach to be set on the fop that was not by
this time yet.
Now, we don't call fop_put_lock() at all. Instead we use ref_put() directly.

Change-Id: I2d8ee3428dcc46a6e217ae0ff10a8dc70bc8ea3c