MERO-1556: ios: fix the few balloc-ed frags case
When the disk space is exhausted or there is fragmentation
balloc groups may not have enough space to serve the
allocation request in one call in one fragment. The old
code did not handle such situation correctly and just
panic-ed.

Now, the code is fixed and we introduce the new constant -
BALLOC_FRAGS_MAX. We do need to know the maximum possible
amount of balloc fragments when we calculate the BE
credits. If balloc is exhausted in such a way that it
can't serve the allocation in less than BALLOC_FRAGS_MAX
fragments - we return ENOSPC error to user.

Also, stio_desc construction was refactored in ios. Now,
we allocate and populate all needed stio_desc structs at
M0_FOPH_INIT phase, so by the time we need to calculate
BE credits we have all needed si_stob index vectors
ready as they will be in actual IOs. This allowed to
remove the dirty hack in stob/ad credits calculations
where we used LNET_MTU to estimate the number of actual
balloc and emap fragments.

Reviewed-by: Hua Huang <hua.huang@seagate.com>
Reviewed-by: Maxim Medved <max.medved@seagate.com>
Reviewed-by: Madhavrao Vemuri <madhav.vemuri@seagate.com>

Change-Id: Ia9d24384d6f00183d62afa15b72e878fe6c8e4ff