m0kdump2trace: doen't rely on mero logs in dmesg
Improve mechanism which is used to extract Mero's kernel-mode trace
buffer from a kernel crashdump. It doesn't rely on the m0mero.ko log
messages with trace-related information to be available in a dmesg
buffer. Instead it's based on the extraction of such information
directly from the dumped kernel memory, locating it by a known magic
identifier.

In addition, there are two new options for m0kdump2trace utility:

  -a|--address <addr>

          In memory address of either a trace buffer or a module image
          (if '-m|--module-image' is provided).

  -s|--size <size>

          Size of the module image, effective only when
          '-m|--module-image' is provided.

Change-Id: I2ef78785e41c36c9fed0b8a8866372abbf4bbdb6