build: unconditionally enable systemd integration by default
Since main target platform for Mero is Centos7 which is based on
systemd, we can safely require systemd's libraries and headers for Mero
build by default. As a side-effect, this ensures a correct Mero build on
other systems that also use systemd, like NixOS.

If it's required to build Mero for non-systemd OS, it's possible to
disable systemd integration using '--disable-systemd' option of Mero
`configure` script.

Change-Id: Ibee3aafa39635b44f85a0038cc7a31e4bb5c7edd