Simplify base directory discovery in m0setup
There were two path before:
- remove a trailing `utils` from the dirname, then add it again
- remove `/usr/bin` or `/usr/sbin` prefix which, hopefully, makes the
directory name empty, and search the path.

There first path amounts to the identity, the second path is rather
dangerous: `m0setup` will malfunction as soon as it is installed in a
non-standard place (say `/user/local/bin`). On the other hand, in
practice, `m0gendisks` (which is the only file for which this directory
discovery is performed) is always installed in the same repository as
`m0setup`.

Change-Id: I005210fb65c8295cc63ad437c87053cf6bb7878f