MERO-1465 Connect to services based on ha state of the service object
During starting of m0d/client, connecting to the offline service causes
failure/stuck of the mod.
To avoid this use ha state in the configuration service object and connect
only if service is online. It assumes that HA will updates the state of the
service object as soon as it is known.
Service context is build for all the services.

Change-Id: I1f89703c5b59875d5cd1ddadcc2d88085b0eaa41