ios: open transactions after network I/O is completed
The idea is to shorten the lifetime of the transaction,
especially in opened state. This allows BE tx group to
accumulate the required number of transactions faster and
to be closed faster. And, as result, to process more
transactions at the same time interval increasing
the overall system throughput.

Reviewed-by: Madhavrao Vemuri <madhav.vemuri@seagate.com>

References: MERO-1613.

Change-Id: If7eb789636822018d5eb642195347acd1df71492