MERO-1673 start mero even if some disks are not available
When some disks specified in configuration are in failed state,
currently mkfs fails and ioservice crashes too.

Now during mkfs -ENOENT error for disk is handled by notifying M0_NC_FAILED to
halon and continue with the initialisation of rest of alive devices.

If linux stob for the failed disk is already created then it is destroyed as
as part of cleanup.

Change-Id: I7b191527df3b4e7e243ccafacf0b3b0f5c841d87