initscripts: Use different endpoints for mkfs and m0d's
When same endpoint is used for mkfs and m0d, some of m0d's may start
earlier and will connect to mkfs instead of corresponding m0d.
Due to which sns repair/rebalance will fail, as some of connections
between m0d's are not available.

To fix this different endpoints are used for mkfs and m0d.

Change-Id: I594a30e8ee8993d8ae41f2c3eaf8eef42ef006ab