MERO-1706 m0d: Use M0_CST_NR to store service contexts
As dummy services ds1, ds2 could be present in conf, which
are not registered. So when a new service type is added after them,
total number of registered services could be less than the value of
this service type. Due to which this service won't be started.

To fix this used M0_CST_NR for storing services instead of count of
registered services.
rc_max_services is not needed, so removed it.

Change-Id: I55dc3956c4f423f3030ee36035a65e8162831f3c