confc: fix the case of process' multi-word core mask
+ Previously b_nr was set to a number of 64-bit words in xcode, while it
should be a number of bits in bitmap.

See confc-ut. There are two processes in the conf.xc. Both of those have
core mask [1: 3] (one 64-bit word of 0x3). Which means, there are 2
significant bits per process. So, per process the count is 2, and 2
processes ultimately result in 4. Previously the bitmap's b_nr was set
to 1 (not 64 as it now is), and only the very first bit was counted per
process.

Change-Id: Ia88b3049441c3f7d1f424f91d2a8f2a7a4af35b1