MERO-1744: Integrate pkg-config into Mero build system
pkg-config is a standard UNIX mechanism to provide necessary compilation/build
information about a package for its users.

As Halon builds against libmero it requires to know correct build flags,
which are currently hard-coded inside Halon build system. That leads to a
potential problem of binary incompatibility between Halon and Mero. To fix it,
Mero will provide a pkg-config information and Halon will use to link with Mero.

Change-Id: I0b66acfdaba6662d18f86394897b514df73ce3b6