cas: add COF_SLANT operation flag
"Slant" flag can be applied to the NEXT operation. It allows iteration
to start with the smallest key following the start key. It's useful when
a user doesn't know exactly the key to start iteration from or if he
wants to start iteration from the first key. In the latter case he
simply provides a zero key.

* [piggyback] Small stylistic fix in cas/service.c

Change-Id: Ib0e21a2513af5f62e54574d0f4a737ec971b5b06