[MERO-1772] st/m0mount-fail: fix the ST and remove cruft
The mount command in the ST used to fail nicely (as expected),
but the finalization routine did something wrong, and the script
exited with non-zero status.

There were some mistakes in the script (e.g., it named the sandbox
directory after another ST) and it did unnecessary things (like
starting second m0d process which was never used).  It also borrowed
m0t1fs/linux_kernel/st/ definitions, which it could do perfectly well
without.

So I've simlified the script.  It still tests the use case, which
it was created for, but gets to the point with far less unnecessary
operations.

Change-Id: I9d2165150858d8d8aae8a1f08fbeed08fec5d837