MERO-1817 22rpc-cancel-test TC.4 failure
After cancelling session while read processes are ongoing in TC.4, to
collect the count of read processes cancelled, it is checked if
processes have failed with the "Input output" error. But in reality,
the read processes may fail even with the "Operation canceled" error.

As a fix, now checking if the read processes have failed either with the
"Input output" error or with the "Operation canceled" error.

Additinally, adding a M0_PRE() at m0_rpc_session_cancel() to verify that
the session being canceled is either in M0_RPC_SESSION_BUSY or in
M0_RPC_SESSION_IDLE state, to confirm that it is a session well established
before being canceled.

Change-Id: Id2ac0c1ffe989803ef594928d02be5901c615701