m0t1fs: fix sporadic EFAULT on copy_from_user()
Disabling pagefaults when making copy_from_user()
could lead to the situations when the user pages
are no longer available (due to kernel's page
reclaiming procedures on memory tension cases).
As result, we could not copy the user data and
generated EFAULT errors in such a cases.

Now, we just do not disable pagefaults on copying
data from userspace. It seems to be safe, since all
the copying is done in context of user process.

Reference: MERO-1295.

Change-Id: I5759ee55035141a6fcacf67504e878c86bc0dfe3