MERO-1884: Fix pool machine state transition crash for failures > K
When number of device failures are greater than K, then pool version becomes
dud due to "i < state->pst_max_device_failures" clients and m0d's are
crashed.

To fix this issue error is logged for this case with message "Pool version is
dud!!" and is handled during repair/rebalance by marking pool version dirty and
repair continues with rest of pool versions. And partial repaired state for pool
is send to halon.

- Also mdpool is skipped while sending status of repair/rebalance to halon.
- pool machine state transit is called during pool version init.

Reviewed-by: Rajanikant Chirmade <rajanikant.chirmade@seagate.com>

Change-Id: Ie576b20302a1e2f55b9e6c05366dea6c97f6cf9b