MERO-1929 Add ST to check M0_DEVICE_DETACH during I/O
The system test sends M0_DEVICE_DETACH command to ioservice when I/O is
in progress. Goal is to check proper stob finalisation with concurrent
users.

There is a race window when an io_fom is finished and new one doesn't
get reference to the storage_dev. In this case M0_DEVICE_DETACH is
performed without a concurrent user and such a scenario is covered by
a spiel system test. To reduce chance of this situation current test
performs $ITER_NR iterations.

Change-Id: Ib948075b5fd5192539dae80ceb2fa3e57fd01bfa