lib: store full path to exec binary in m0_argv0 variable
When analyzing core dumps, it may be important to know the exact path
to the `m0d` binary which has crashed. Unfortunately, when core file is
created, it stores the information about process' command line provided
by ptrace API. Which makes the hack of replacing argv[0] by
/proc/self/exe content ineffective.

To workaround this problem, libmero.so stores a full path of the
executable in 'm0_argv0' global variable, which can be inspected in a
debugger. Also, initscripts launch `m0d` and `m0mkfs` binaries using a
full path instead of just a name.

Change-Id: Ifb8e80e1fa13a3eda16243c2f0326365484b4ab0