rpc: move HA timer from item to connection
To reduce the number of notifications, HA timer was moved from a separate rpc
item to a connection. Timer starts (if it was not) when item_resend() or
item_timedout() occurs. When it is triggered, Mero sends a HA message with a
service HA state only once. In case when the service is M0_NC_TRANSIENT, and
reply for an item was received, Mero notifies HA about M0_NC_ONLINE state.

The interval for HA timer is configurable and may be changed with
m0_rpc_conn_ha_cfg_set() for a separate connection.

Jira: MERO-1813

Change-Id: If7e69f601b616cf3296330fca82b549ac150557a