utils/m0trace: fix incorrect file path stripping
m0trace tries to show file paths relative to the mero top build dir.
This fixes a bug in the regular expression used for stripping common
prefix path, which causes 'mero' subdirectory inside main 'mero'
directory to be stripped as well. So, for example, files like
'mero/ha.c' appeared in the logs as 'ha.c'.

Change-Id: I822271096a9fbce1900205294bc8284ea3cc0955