be: fix m0_be_btree_insert_credit/2()
The number of required credits calculated by
m0_btree_insert_credit/2() routines was less than
required. The problem was revealed by using
m0_be_btree_insert_credit2() in UTs.

m0_be_btree_insert_credit2() is more accurate as it
uses the current btree height to calculate the credits,
and is more likely to fail in case of any bug in
credits calculation. So we better use it in the UTs.

References: MERO-1948.

Change-Id: Ifa569cf701c579786282717c89d40a664e0b6f9a