MERO-1960 RPC: Fix connection to offline services in case of restart
When cluster is restarted, service offline events are send to halon
during m0d shutdown. While restarting of m0d during the creatiion of the
service contexts connection to services get cancelled as they are offline
as per halon and currenlt crashes in rpc item replied call back as it
assumes rpc post is always successful.

This crash is fixed by adding the case of connection could be in initialised
state.
So initially connection is in failed state and when services become online,
connection context is rebuild in service event handlers upon receiving
the online events from halon.

Change-Id: I69d9296486a8335fafee6a0dec5a998e5bba8179