rpc: remove _0C macros from precondition
The macros is used with conjunction but not with disjunction.

M0_PRE(_0C(!m0_fid_is_set(&conn->c_svc_fid)) ||
       _0C(m0_fid_eq(&conn->c_svc_fid, svc_fid)));

In case when the first is false but the second is true, m0_failed_condition will
be equal to the first, but in fact there is no failed condition.

Change-Id: I2d7af40679f217e9c3e701c867331dc26d967df6