stob: disk error emulation introduced
Now one can attach to m0d process with gdb and
set emulate_disk_errors_nr global variable to
some value, like 1000 to emulate a number of
EIO disk errors on some disk. The bigger value -
the more errors will be generated for the disk
until it exists (i.e. not detached yet).

When the emulate_disk_errors_nr value is dropped
to zero - we still remember the last disk fid. On
the next try to emulate the errors we will try to
find the same disk. And only if it was not found
during the all emulate_disk_errors_nr tries -
we will forget that disk. On the next try another
disk will be chosen. This algorithm alows us to
emulate errors to the same disk even if it was
reset by the halon and attached back again.

Change-Id: Iadacfa2936d24ff70dadbfee208cdcf4d1c8c059