MERO-2001 rconfc: wait until all confc contexts become complete on stopping
If there are multiple confd instances in the cluster, it is possible that rconfc
is reached quorum and elected a version. But requests to some confds are still
in progress and corresponding confc contexts in S_WAIT_REPLY state. In this case
rpc sessions should be cancelled on rconfc stopping and rconfc has to wait until
all contexts become to a terminal state. As soon as it happened, contexts may be
finalised correctly.

Other changes:
* Fake HA returns all confd instances in m0_ha_entrypoint_rep, not the first
  founded.
* Test for the case of multiple confd was added.

Change-Id: I6b395bff693dd19f56041f37a42ee9f5d723cb0d