MERO-1985 Fix deadlock on storage device attach
m0_storage_dev_attach() creates a synchronous BE tx implicitly during
balloc initialisation. This is done under m0_storage_devs lock. There
can a situation when BE tx of an I/O fom is grouped to the same tx
group. In this case the group is closed only when both transactions are
closed. But the fom needs to acquire m0_storage_devs lock to close the
second tx.
We have deadlock when first thread needs to close a tx synchronously
before releasing a lock and second thread needs to acquire the lock to
let the tx group be closed.

Fix this situation by moving storage device initialisation part outside
m0_storage_devs lock.

Change-Id: Idecdac5ea285cf0f28e529cca3c2578bd2c7797e