stob/ioq: don't acquire a reference to a stob on I/O error
ioq_io_error() acquires a reference to a stob on the error path. This is
done in order to ensure that the stob exists when it's used in an AST.
This behaviour is not synchronised with m0_storage_dev_release() which
destroys a stob on releasing the last reference. As result m0d crashes
when spiel sends device_detach command to an ioservice with an active
I/O to the device and the I/O fails.

Fix the above problem by removing the AST and moving its code to
ioq_io_error(). We may do this, because the AST doesn't perform blocking
operations.

References: MERO-1985
Change-Id: If14f0c552d69a939b57c059d5339196f0fb5ba54