MERO-2015 m0t1fs: Newly created files get the pool version based on RECD
Currently while creating the file pool version dirty flag is checked
before traversing the configuration.
But due to this even if device or controller is back online, these are
not used until current pool version becomes dirty.
To avoid this checking for dirty flag is removed and pool versions are
build on RECD events and current pool version is stored in pools common.

Also during mount if no valid pool version is found, instead of failing,
pool version is not assigned to root inode.

Change-Id: Ic9d694b2f123c3103edca3d4c521d4861e36ca5d