be/btree: support longer values during insertion with overwrite flag and update
Also add tests for insertion with overwrite flag and updating using longer values.

If the length of old value is greater than or equal to the length of new value
then old value area is reused for new value and no alloc/free operations are
done.
If the length of old value is less than the length of new value then update is
done as following:
  1. delete old key/value;
  2. insert new key/value.

To consider credits for possible deletion and insertion:
  - In case of insertion with overwrite flag the caller must always calculate
    delete credits along with insert credits.
  - In case of updating the function m0_be_btree_update_credit() should be
    called if data has fixed length and existing value area can be re-used
    for a new value (no alloc/free operations are needed). Otherwise the function
    m0_be_btree_update_credit2() should be called, it considers credits for
    possible deletion and insertion.

Change-Id: If07c41bf667f20fed3ae55b4064315c965475bcd