MERO-2120 cob: Add missing create credits in case of recreation of mdcob
When creating a mdcob, if it already exists in cob domain but with a
different pool version then it is deleted and cob with new pool version
is created. It is found that credits for create are not added in
recreation case, which causes panic during  be_btree_insert().
Fixed this by adding create credits for mdcob when recreate is true.

Change-Id: Ic26dc3bfd0b980ca7fb81a9bd1d3018cab087820