MERO-2150 rconfc: repeat querying entrypoint on version election failure
The situation when quorum appears not reached on the currently
configured confd set should result in querying entrypoint information
from Halon and retrying version election. The process should repeat
until succeeded.

NB: The design is based on the assumption that Halon is able to
detect critical issues related to configuration and accommodate
them properly.

Change-Id: Ib8735337ac2d1290d6847b9d85c67780a62f905f