build: make check-public-api more strict
Remove name filter for exported function names in check-public-api
script. It turns out to be more practical to catch any exported
function from libmero.so and not only those starting with 'm0_' prefix.

This will help to avoid situations when accidentally exported function
with a common name like 'crc32' can lead to a crash, because of a
conflict with a third party library (like 'zlib') which uses a func with
the same name. See 955a05d8c6822b 'file/crc: make crc32() function static'.

Change-Id: I343bf87eca8c5acde803c62762f9e5bdf3d240aa