build: use correct versioning scheme for libmero.so
Allow external users of libmero.so to link with it using only major
version number:

  /usr/lib64/libmero.so -> libmero.so.1.0.0
  /usr/lib64/libmero.so.1 -> libmero.so.1.0.0
  /usr/lib64/libmero.so.1.0.0

By default, when application is linking with libmero it will choose
highest major version available in the system. In this example it's
'libmero.so.1'. If more specific version is required, it can be
specified with linker option, like this:

  gcc -o app app.o -lmero.so.1.0.0

When newer version of the library is installed, there is no need rebuild
the application, as long as library follows semantic versioning logic
and doesn't break API w/o updating major version number.

Change-Id: Ifc12513c4f33cbe5c05805821140fac809985dfc