MERO-2284 Wait for completion of asynchronous connection to the services
During mount all remote services are connected asynchronously.
And if any item is sent before these connections are created it will
fail with session invalid errors.
To fix this issue wait until completion of connection to remote online
services and then send rpc items to them.
It is done for clients only, as io could be done immediately.

Change-Id: I9b5c6294250239ed4f199c0bf5a407c0aeb10772