MERO-2281 rconfc: do not disconnect confc synchronously in AST
Disconnecting synchronously in AST blocks rconfc sm group for the time
of disconnection. The same sm group is used in m0_rconfc::rc_confc. This
way the confc cache appears locked for the time blocking the other
threads trying to access conf objects in the cache.

Infinite disconnection timeout may cause a deadlock.

Change-Id: Ib869e614926f1026c683a143772af2da088ffe9f